refactor(cli): split command modules
Keep the public CLI runner in lib.rs while moving clap args, command handlers, shell integration, output, environment, and errors into focused modules.
