+++ crates/cli/src/commands/php.rs
@@ -1,0 +1,13 @@
+use std::process::ExitCode;
+
+use crate::args::PhpInstallArgs;
+use crate::error::{CliError, ExecuteError};
+
+pub(crate) fn install(args: PhpInstallArgs) -> Result<ExitCode, ExecuteError> {
+    let _track = args.track;
+
+    Err(CliError::DeferredCommand {
+        command: "php:install",
+    }
+    .into())
+}
